// src/modules/product-adm/facade/product-adm.facade.interface.ts
export default interface ProductAdmFacadaInterface {

    addProduct(input: AddProductFacadeInputDTO): Promise<void>
    checkStock(input: CheckStockFacadeInputDTO): Promise<CheckStockFacadeOutputDTO>

}

export interface AddProductFacadeInputDTO {
    id?: string
    name: string
    description: string
    purchasePrice: number
    stock: number
}

export interface CheckStockFacadeInputDTO {
    productId: string
}

export interface CheckStockFacadeOutputDTO {
    productId: string
    stock: number
}